refactor(ts-ioc-container): extract getConstructorChain into utils
The prototype-chain walk was a private helper in hook.ts, though nothing about
it is hook-specific: reading own metadata off each class in turn is how any
metadata reader inherits what ancestors declared, since Reflect.getOwnMetadata
never looks up the chain itself. It moves to lib/utils/getConstructorChain.ts
and is exported alongside the other utils, with its own tests.

No behavior change: getHooks and hasHooks call the same function they did.
